requirements.txt
Django==1.8.3
Markdown==2.6.2
Pygments==2.0.2
django-filter==0.10.0
djangorestframework==3.1.3
httpie==0.9.2
pytz==2015.4
requests==2.7.0
wsgiref==0.1.2
